Backend integrado ao Frontend

+++ src/main/java/com/surittec/crud/controller/ClienteController.java
@@ -5,15 +5,19 @@ import java.util.List;
 import javax.validation.Valid;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.surittec.crud.entity.Cliente;
+import com.surittec.crud.repository.ClienteRepository;
 import com.surittec.crud.service.ClienteService;
 
 @RestController
@@ -22,8 +26,12 @@ public class ClienteController {
 	@Autowired
 	private ClienteService clienteService;
 	
+	@Autowired
+	private ClienteRepository clienteRepository;
+	
 	@PostMapping("/clientes")
-	public Cliente addCliente(@RequestBody @Valid Cliente Cliente) {
+	@ResponseStatus(HttpStatus.CREATED)
+	public Cliente addCliente(@Valid @RequestBody Cliente Cliente) {
 		return clienteService.saveCliente(Cliente);
 	}
 	
@@ -38,12 +46,23 @@ public class ClienteController {
 	}
 	
 	@PutMapping("/clientes/{id}")
-	public Cliente updateCliente(@RequestBody Cliente Cliente, @PathVariable Long id) {
-		return clienteService.updateCliente(Cliente);
+	public ResponseEntity<Cliente> updateCliente(@Valid @PathVariable Long id, @RequestBody Cliente cliente) {
+		if(!clienteRepository.existsById(id)) {
+			return ResponseEntity.notFound().build();
+		}
+		
+		cliente.setId(id);
+		cliente = clienteService.saveCliente(cliente);
+		return ResponseEntity.ok(cliente);
 	}
 	
 	@DeleteMapping("/clientes/{id}")
-	public String deleteCliente(@PathVariable Long id) {
-		return clienteService.deleteCliente(id);
+	public ResponseEntity<Void> deleteCliente(@PathVariable Long id) {
+		if(!clienteRepository.existsById(id)) {
+			return ResponseEntity.notFound().build();
+		}
+		
+		clienteService.deleteCliente(id);
+		return ResponseEntity.noContent().build();
 	}
 }
